clean up

+++ src/stuff.cpp
@@ -25,7 +25,7 @@ bool getEspShallRestart(){
 String epochToTimeString(unsigned long offset){
     time_t t = startTime + offset / 1000;
     char timestring[30];
-    sprintf(timestring, "%4i-%2i-%2i-%2i-%2i-%2i-%4i", year(t),month(t),day(t),hour(t),minute(t),second(t),offset%1000);
+    sprintf(timestring, "%4i-%2i-%2i-%2i-%2i-%2i-%4i", (int) year(t),month(t),day(t),hour(t),minute(t),second(t),offset%1000);
     return timestring;
 }
